Move test runner in a lib crate
